COMP: Separate each derived class into file
Each derived class is placed in it's own file.
This allows for comparison of different child class
types directly, and facilitates code maintenance.

This separation has a side benefit of facilitating
debugging by providing a smaller scope for review.

Provided a consistent code formatting to minimize
differences between files.
