+++ ukf/filter_model.h
@@ -192,550 +192,8 @@ inline mat33_t SetIdentityScaled(double diff_fw)
    return tmp;
 }
 
-/**
- * \struct Full1T_FW
- * \brief Full 1-Tensor model with free water
- *
- * Model describing 1-tensor tractography with the full tensor representation (3 angles, 3 eigenvalues)
- * and free water estimation.
-*/
-class Full1T_FW : public FilterModel
-  {
-public:
-  Full1T_FW(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType qw, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained,
-            const ukfPrecisionType diff_fw)
-    : FilterModel(7, rs, weights_on_tensors, constrained), _lambda_min(100.0), m_D_iso(SetIdentityScaled(diff_fw))
-  {
-#if 0
-    m_D_iso << diff_fw, 0, 0,
-      0, diff_fw, 0,
-      0, 0, diff_fw;
-#endif
+/* Common Function used for NODDI */
+extern void createProtocol(const ukfVectorType& _b_values,
+                    ukfVectorType& _gradientStrength, ukfVectorType& _pulseSeparation);
 
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = qs;
-    _Q(3, 3) = _Q(4, 4) = _Q(5, 5) = ql;
-    _Q(6, 6) = qw;
-
-    _D.resize(7, 5);
-    _D.setConstant(ukfZero);
-
-    _d.resize(5);
-
-    // Setting the constraints according to D'*x >= -d
-    _D(6, 0) = -1;
-    _d(0) = 1;  // w <= 1
-    _D(6, 1) = 1;
-    _d(1) = 0;  // w >= 0
-    _D(3, 2) = 1;
-    _d(2) = 0;  // l1 >= 0
-    _D(4, 3) = 1;
-    _d(3) = 0;  // l2 >= 0
-    _D(5, 4) = 1;
-    _d(4) = 0;  // l3 >= 0
-
-  }
-
-  virtual ~Full1T_FW()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor1T(const State& x, vec3_t& m, vec3_t& l);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-
-  /** apparent diffusion coefficient of free water */
-  mat33_t m_D_iso;
-
-  };
-
-/**
- * \struct Full1T
- * \brief Full 1-Tensor model without free water
- *
- * Model describing 1-tensor tractography with the full tensor representation
- * (3 angles, 3 eigenvalues).
-*/
-class Full1T : public FilterModel
-  {
-public:
-  Full1T(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained)
-    : FilterModel(6, rs, weights_on_tensors, constrained), _lambda_min(100.0)
-  {
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = qs;
-    _Q(3, 3) = _Q(4, 4) = _Q(5, 5) = ql;
-  }
-
-  virtual ~Full1T()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor1T(const State& x, vec3_t& m, vec3_t& l);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-  };
-
-/**
- * \struct Full2T
- * \brief Full 2-Tensor model without free water
- *
- * Model describing 2-tensor tractography with the full tensor representation
- * (3 angles, 3 eigenvalues).
-*/
-class Full2T : public FilterModel
-  {
-public:
-  Full2T(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained)
-    : FilterModel(12, rs, weights_on_tensors, constrained), _lambda_min(100.0)
-  {
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = _Q(6, 6) = _Q(7, 7) = _Q(8, 8) = qs;
-    _Q(3, 3) = _Q(4, 4) = _Q(5, 5) = _Q(9, 9) = _Q(10, 10) = _Q(11, 11) = ql;
-  }
-
-  virtual ~Full2T()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor2T(const State& x, const vec3_t& old_m, vec3_t& m1, vec3_t& l1, vec3_t& m2, vec3_t& l2);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-  };
-
-/**
- * \struct Full2T_FW
- * \brief Full 2-Tensor model with free water
- *
- * Model describing 2-tensor tractography with the full tensor representation (3 angles, 3 eigenvalues)
- * and free water estimation.
-*/
-class Full2T_FW : public FilterModel
-  {
-public:
-  Full2T_FW(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType qw, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained,
-            const ukfPrecisionType diff_fw)
-    : FilterModel(13, rs, weights_on_tensors, constrained), _lambda_min(100.0), m_D_iso(SetIdentityScaled(diff_fw))
-  {
-#if 0
-    m_D_iso << diff_fw, 0, 0,
-      0, diff_fw, 0,
-      0, 0, diff_fw;
-#endif
-
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = _Q(6, 6) = _Q(7, 7) = _Q(8, 8) = qs;
-    _Q(3, 3) = _Q(4, 4) = _Q(5, 5) = _Q(9, 9) = _Q(10, 10) = _Q(11, 11) = ql;
-    _Q(12, 12) = qw; // noise for weights
-
-    _D.resize(13, 8);
-    _D.setConstant(ukfZero);
-
-    _d.resize(8);
-
-    // Setting the constraints according to D'*x >= -d
-    _D(12, 0) = -1;
-    _d(0) = 1;  // w <= 1
-    _D(12, 1) = 1;
-    _d(1) = 0;  // w >= 0
-    _D(3, 2)  = 1;
-    _d(2) = 0;  // l11 >= 0
-    _D(4, 3)  = 1;
-    _d(3) = 0;  // l21 >= 0
-    _D(5, 4)  = 1;
-    _d(4) = 0;  // l31 >= 0
-    _D(9, 5)  = 1;
-    _d(5) = 0;  // l12 >= 0
-    _D(10, 6) = 1;
-    _d(6) = 0;  // l22 >= 0
-    _D(11, 7) = 1;
-    _d(7) = 0;  // l32 >= 0
-  }
-
-  virtual ~Full2T_FW()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor2T(const State& x, const vec3_t& old_m, vec3_t& m1, vec3_t& l1, vec3_t& m2, vec3_t& l2);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-
-  /** apparent diffusion coefficient of free water */
-  const mat33_t m_D_iso;
-  };
-
-/**
- * \struct Full3T
- * \brief Full 3-Tensor model with free water
- *
- * Model describing 3-tensor tractography with the full tensor representation
- * (3 angles, 3 eigenvalues).
-*/
-class Full3T : public FilterModel
-  {
-public:
-  Full3T(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained)
-    : FilterModel(18, rs, weights_on_tensors, constrained), _lambda_min(100.0)
-  {
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = qs;
-    _Q(6, 6) = _Q(7, 7) = _Q(8, 8) = qs;
-    _Q(12, 12) = _Q(13, 13) = _Q(14, 14) = qs;
-
-    _Q(3, 3) = _Q(4, 4) = _Q(5, 5) = ql;
-    _Q(9, 9) = _Q(10, 10) = _Q(11, 11) = ql;
-    _Q(15, 15) = _Q(16, 16) = _Q(17, 17) = ql;
-  }
-
-  virtual ~Full3T()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor3T(const State& x, const vec3_t& old_m, vec3_t& m1, vec3_t& l1, vec3_t& m2, vec3_t& l2, vec3_t& m3,
-                            vec3_t& l3);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-
-  };
-
-/**
- * \struct Simple1T_FW
- * \brief Simple 1-Tensor model with free water
- *
- * Model describing 1-tensor tractography with the simplified tensor representation (two minor eigenvalues are equal)
- * and free water estimation
-*/
-class Simple1T_FW : public FilterModel
-  {
-public:
-  Simple1T_FW(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType qw, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors,
-              bool constrained, const ukfPrecisionType diff_fw)
-    : FilterModel(6, rs, weights_on_tensors, constrained), _lambda_min(100.0), m_D_iso(SetIdentityScaled(diff_fw))
-  {
-#if 0
-    m_D_iso << diff_fw, 0, 0,
-      0, diff_fw, 0,
-      0, 0, diff_fw;
-#endif
-
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = qs;
-    _Q(3, 3) = _Q(4, 4) = ql;
-    _Q(5, 5) = qw; // noise for weights
-
-    _D.resize(6, 4);
-    _D.setConstant(ukfZero);
-
-    _d.resize(4);
-
-    // Setting the constraints according to D'*x >= -d
-    _D(5, 0) = -1;
-    _d(0) = 1;  // w <= 1
-    _D(5, 1) = 1;
-    _d(1) = 0;  // w >= 0
-    _D(3, 2) = 1;
-    _d(2) = 0;  // l1 >= 0
-    _D(4, 3) = 1;
-    _d(3) = 0;  // l2 >= 0
-
-  }
-
-  virtual ~Simple1T_FW()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor1T(const State& x, vec3_t& m, vec3_t& l);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-
-  /** apparent diffusion coefficient of free water */
-  mat33_t m_D_iso;
-
-  };
-
-/**
- * \struct Simple1T
- * \brief Simple 1-Tensor model
- *
- * Model describing 1-tensor tractography with the simplified tensor representation (two minor eigenvalues are equal)
-*/
-class Simple1T : public FilterModel
-  {
-public:
-  Simple1T(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained)
-    : FilterModel(5, rs, weights_on_tensors, constrained), _lambda_min(100.0)
-  {
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = qs;
-    _Q(3, 3) = _Q(4, 4) = ql;
-  }
-
-  virtual ~Simple1T()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor1T(const State& x, vec3_t& m, vec3_t& l);
-
-  const ukfPrecisionType _lambda_min;
-  };
-
-/**
- * \struct Simple2T
- * \brief Simple 2-Tensor model
- *
- * Model describing 2-tensor tractography with the simplified tensor representation (two minor eigenvalues are equal).
-*/
-class Simple2T : public FilterModel
-  {
-public:
-  Simple2T(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained)
-    : FilterModel(10, rs, weights_on_tensors, constrained), _lambda_min(100.0)
-  {
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = _Q(5, 5) = _Q(6, 6) = _Q(7, 7) = qs;
-    _Q(3, 3) = _Q(4, 4) = _Q(8, 8) = _Q(9, 9) = ql;
-  }
-
-  virtual ~Simple2T()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor2T(const State& x, const vec3_t& old_m, vec3_t& m1, vec3_t& l1, vec3_t& m2, vec3_t& l2);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-
-  };
-
-/**
- * \struct Simple2T_FW
- * \brief Simple 2-Tensor model with free water
- *
- * Model describing 2-tensor tractography with the simplified tensor representation (two minor eigenvalues are equal)
- * and free water estimation
-*/
-class Simple2T_FW : public FilterModel
-  {
-public:
-  Simple2T_FW(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType qw, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors,
-              bool constrained, const ukfPrecisionType diff_fw)
-    : FilterModel(11, rs, weights_on_tensors, constrained), _lambda_min(100.0), m_D_iso(SetIdentityScaled(diff_fw))
-  {
-#if 0
-    m_D_iso << diff_fw, 0, 0,
-      0, diff_fw, 0,
-      0, 0, diff_fw;
-#endif
-
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = _Q(5, 5) = _Q(6, 6) = _Q(7, 7) = qs;
-    _Q(3, 3) = _Q(4, 4) = _Q(8, 8) = _Q(9, 9) = ql;
-    _Q(10, 10) = qw; // noise for weights
-
-    _D.resize(11, 6);
-    _D.setConstant(ukfZero);
-
-    _d.resize(6);
-
-    // Setting the constraints according to D'*x >= -d
-    _D(10, 0) = -1;
-    _d(0) = 1;  // w <= 1
-    _D(10, 1) = 1;
-    _d(1) = 0;  // w >= 0
-    _D(3, 2)  = 1;
-    _d(2) = 0;  // l11 >= 0
-    _D(4, 3)  = 1;
-    _d(3) = 0;  // l21 >= 0
-    _D(8, 4)  = 1;
-    _d(4) = 0;  // l12 >= 0
-    _D(9, 5)  = 1;
-    _d(5) = 0;  // l22 >= 0
-  }
-
-  virtual ~Simple2T_FW()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor2T(const State& x, const vec3_t& old_m, vec3_t& m1, vec3_t& l1, vec3_t& m2, vec3_t& l2);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-
-  /** apparent diffusion coefficient of free water */
-  mat33_t m_D_iso;
-  };
-
-/**
- * \struct Simple3T
- * \brief Simple 3-Tensor model
- *
- * Model describing 3-tensor tractography with the simplified tensor representation (two minor eigenvalues are equal)
-*/
-class Simple3T : public FilterModel
-  {
-public:
-  Simple3T(ukfPrecisionType qs, ukfPrecisionType ql, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained)
-    : FilterModel(15, rs, weights_on_tensors, constrained), _lambda_min(100.0)
-  {
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = qs;
-    _Q(5, 5) = _Q(6, 6) = _Q(7, 7) = qs;
-    _Q(10, 10) = _Q(11, 11) = _Q(12, 12) = qs;
-
-    _Q(3, 3) = _Q(4, 4) = _Q(8, 8) = _Q(9, 9) = _Q(13, 13) = _Q(14, 14) = ql;
-  }
-
-  virtual ~Simple3T()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  virtual void State2Tensor3T(const State& x, const vec3_t& old_m, vec3_t& m1, vec3_t& l1, vec3_t& m2, vec3_t& l2, vec3_t& m3,
-                            vec3_t& l3);
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-
-  };
-
-/**
- * \struct NODDI1F
- * \brief NOODI 1 fiber model
- *
- * Model describing 1-fiber tractography with the NODDI representation
-*/
-class NODDI1F : public FilterModel
-  {
-public:
-  NODDI1F(ukfPrecisionType qs, ukfPrecisionType qkappa, ukfPrecisionType qvic, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained)
-    : FilterModel(6, rs, weights_on_tensors, constrained), _lambda_min(100.0)
-  {
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = qs;
-     _Q(4, 4) =qkappa;
-     _Q(3, 3) = _Q(5,5)= qvic;
-
-    _D.resize(6, 5);
-    _D.setConstant(ukfZero);
-
-    _d.resize(5);
-
-    // Setting the constraints according to D'*x >= -d
-    _D(5, 0) = -1;
-    _d(0) = 1;  // Viso <= 1
-    _D(5, 1) = 1;
-    _d(1) = 0;  // Viso >= 0
-
-    _D(3, 2) = 1;
-    _d(2) = 0;  // Vic >= 0
-    _D(3, 3) = -1;
-    _d(3) = 1;  // Vic <= 1
-
-    _D(4, 4) = 1;
-    _d(4) = 0;  // kappa >= 0
-     
-  }
-
-  virtual ~NODDI1F()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  const ukfPrecisionType _lambda_min;
-  };
-
- /**
- * \struct Simple2T
- * \brief Simple 2-Tensor model
- *
- * Model describing 2-tensor tractography with the simplified tensor representation (two minor eigenvalues are equal).
-*/
-class NODDI2F : public FilterModel
-  {
-public:
-  NODDI2F(ukfPrecisionType qs, ukfPrecisionType qkappa, ukfPrecisionType qvic, ukfPrecisionType rs, const ukfVectorType& weights_on_tensors, bool constrained)
-    : FilterModel(11, rs, weights_on_tensors, constrained), _lambda_min(100.0) 
-  {
-    _Q(0, 0) = _Q(1, 1) = _Q(2, 2) = _Q(5, 5) = _Q(6, 6) = _Q(7, 7) = qs;
-    _Q(4, 4)= _Q(9, 9) = qkappa;
-    _Q(3, 3) = _Q(8, 8) = _Q(10, 10) = qvic; // noise for weights
- 
-    _D.resize(11, 8);
-    _D.setConstant(ukfZero);
-
-    _d.resize(8);
-
-    // Setting the constraints according to D'*x >= -d
-    _D(10, 0) = -1;
-    _d(0) = 1;  // Viso <= 1
-    _D(10, 1) = 1;
-    _d(1) = 0;  // Viso >= 0
-
-    _D(3, 2) = 1;
-    _d(2) = 0;  // Vic >= 0
-    _D(3, 3) = -1;
-    _d(3) = 1;  // Vic <= 1
-
-    _D(4, 4) = 1;
-    _d(4) = 0;  // kappa >= 0
-
-    _D(8, 5) = 1;
-    _d(5) = 0;  // Vic >= 0
-    _D(8, 6) = -1;
-    _d(6) = 1;  // Vic <= 1
-
-    _D(9, 7) = 1;
-    _d(7) = 0;  // kappa >= 0
-
-  }
-
-  virtual ~NODDI2F()
-  {
-  }
-
-  virtual void F(ukfMatrixType& X) const;
-
-  virtual void H(const  ukfMatrixType& X, ukfMatrixType& Y) const;
-
-  /** The minimum value of the eigenvalues. Clamped in each step */
-  const ukfPrecisionType _lambda_min;
-
-  };
 #endif  // FILTER_MODEL_H_
